Fix shift exponent too large UB in CPU

+++ src/hardware/cpu/executor.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015, 2016  Marco Bortolin
+ * Copyright (C) 2015-2021  Marco Bortolin
  *
  * This file is part of IBMulator.
  *
@@ -203,8 +203,8 @@ public:
 	uint16_t AND_w(uint16_t op1, uint16_t op2);
 	uint32_t AND_d(uint32_t op1, uint32_t op2);
 
-	void BT_w(uint16_t _op1, uint16_t _op2);
-	void BT_d(uint32_t _op1, uint32_t _op2);
+	void BT_w(uint32_t _op1, uint16_t _op2);
+	void BT_d(uint64_t _op1, uint32_t _op2);
 	uint16_t BT_ew(uint16_t _op2, bool _rmw);
 	uint32_t BT_ed(uint32_t _op2, bool _rmw);
